Finalisation du projet, et passage en environnement de production

+++ src/Controller/FigureController.php
@@ -175,7 +175,7 @@ class FigureController extends AbstractController
      */
     public function imageUpdate(Request $request, Image $image): Response
     {
-        if (($this->isCsrfTokenValid('update_image' . $image->getId(), $request->get('_token')))) {
+        if ( ($this->isCsrfTokenValid('update_image' . $image->getId(), $request->get('_token'))) && $request->files->get('image') && $request->request->get('old_image')) {
             $upload_file = $this->getParameter('upload_directory');
             $file_name = $request->request->get('old_image');
 
@@ -196,7 +196,7 @@ class FigureController extends AbstractController
      */
     public function imageDelete(Request $request, Image $image): Response
     {
-        if (($this->isCsrfTokenValid('delete_image' . $image->getId(), $request->get('_token')))) {
+        if (($this->isCsrfTokenValid('delete_image' . $image->getId(), $request->get('_token'))) && $image) {
             $em = $this->getDoctrine()->getManager();
             if (file_exists('uploads/'.$image->getImage()))
                 unlink ('uploads/'.$image->getImage());
@@ -224,7 +224,7 @@ class FigureController extends AbstractController
      */
     public function videoUpdate(Request $request, Video $video): Response
     {
-        if (($this->isCsrfTokenValid('update_video' . $video->getId(), $request->get('_token')))) {
+        if (($this->isCsrfTokenValid('update_video' . $video->getId(), $request->get('_token'))) && $request->request->get('video')) {
             $video->setVideo($request->request->get('video'));
             $em = $this->getDoctrine()->getManager();
 
@@ -243,7 +243,7 @@ class FigureController extends AbstractController
      */
     public function videoDelete(Request $request, Video $video): Response
     {
-        if (($this->isCsrfTokenValid('delete_video' . $video->getId(), $request->get('_token')))) {
+        if (($this->isCsrfTokenValid('delete_video' . $video->getId(), $request->get('_token'))) && $video) {
             $em = $this->getDoctrine()->getManager();
 
             $em->remove($video);
@@ -260,7 +260,7 @@ class FigureController extends AbstractController
      */
     public function imagefeatured(Request $request, Image $image): Response
     { 
-        if (($this->isCsrfTokenValid('featured_image' . $image->getId(), $request->get('_token')))) {
+        if (($this->isCsrfTokenValid('featured_image' . $image->getId(), $request->get('_token'))) && $image->getId()) {
             $figure = $image->getIdFigure();
             $figure->setFeaturedImage($image->getId());
             $em = $this->getDoctrine()->getManager();
